pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentimize"
version = "0.1.0"
description = "Automatic AI agent cost optimizer"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "typer>=0.12",
    "rich>=13",
    "openai>=1.30",
    "networkx>=3.2",
    "pydantic>=2.5",
    "scipy>=1.12",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.scripts]
agentimize = "agentimize.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentimize*"]
